test(wasm): share browser smoke cases between client and parity gate
The Worker smoke sources and their verify options were duplicated
between web/client.mjs and test-browser.mjs, so the native parity gate
could silently drift from what the browser probe actually ran. Move the
cases into web/cases.mjs, drive both probes from it, and carry the
exact verify options per case.
